% Auto-generated: do not edit by hand
\name{baysanChart}

\alias{baysanChart}

\title{BaysanChart component}

\description{
BaysanChart is an example of using Ant Design Charts component. It takes a JSON data to visualize.
}

\usage{
baysanChart(id=NULL, data=NULL, legendPosition=NULL, xField=NULL,
yField=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Unnamed list. The data used to visualize}

\item{legendPosition}{Character. The legendPosition to set the position of the chart legend}

\item{xField}{Character. The xField to use on the x-axis}

\item{yField}{Character. The yField to use on the y-axis}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

